distributed_tool_based_multi_agent/news_agent_service/requirements.txt
langchain==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.31
# langchain-community==0.3.29

langgraph==0.6.5

openai==1.102.0

fastapi==0.116.1
uvicorn==0.35.0

pydantic==2.11.7

httpx==0.28.1
requests==2.32.5

python-dotenv==1.1.1

typing-extensions==4.14.1

tiktoken==0.11.0
